+++ sdk/search/azure-search-documents/azure/search/documents/models/__init__.py
@@ -35,11 +35,13 @@ from .._generated.models import (
     QueryLanguage,
     QuerySpellerType,
     QueryType,
+    RawVectorQuery,
     ScoringStatistics,
     SearchMode,
     SemanticErrorHandling,
     QueryDebugMode,
-    Vector,
+    VectorQuery,
+    VectorQueryKind,
 )
 from .._utils import odata
 
@@ -56,9 +58,11 @@ __all__ = (
     "QueryLanguage",
     "QuerySpellerType",
     "QueryType",
+    "RawVectorQuery",
     "ScoringStatistics",
     "SearchMode",
     "SemanticErrorHandling",
     "QueryDebugMode",
-    "Vector",
+    "VectorQuery",
+    "VectorQueryKind",
 )