+++ sdk/keyvault/azure-mgmt-keyvault/azure/mgmt/keyvault/v2023_02_01/models/_models_py3.py
@@ -82,7 +82,8 @@ class AccessPolicyEntry(_serialization.Model):
 class Action(_serialization.Model):
     """Action.
 
-    :ivar type: The type of action. Known values are: "rotate" and "notify".
+    :ivar type: The type of the action. The value should be compared case-insensitively. Known
+     values are: "Rotate" and "Notify".
     :vartype type: str or ~azure.mgmt.keyvault.v2023_02_01.models.KeyRotationPolicyActionType
     """
 
@@ -94,7 +95,8 @@ class Action(_serialization.Model):
         self, *, type: Optional[Union[str, "_models.KeyRotationPolicyActionType"]] = None, **kwargs: Any
     ) -> None:
         """
-        :keyword type: The type of action. Known values are: "rotate" and "notify".
+        :keyword type: The type of the action. The value should be compared case-insensitively. Known
+         values are: "Rotate" and "Notify".
         :paramtype type: str or ~azure.mgmt.keyvault.v2023_02_01.models.KeyRotationPolicyActionType
         """
         super().__init__(**kwargs)
@@ -1259,7 +1261,8 @@ class ManagedHsm(ManagedHsmResource):
 class ManagedHsmAction(_serialization.Model):
     """ManagedHsmAction.
 
-    :ivar type: The type of action. Known values are: "rotate" and "notify".
+    :ivar type: The type of the action. The value should be compared case-insensitively. Known
+     values are: "Rotate" and "Notify".
     :vartype type: str or ~azure.mgmt.keyvault.v2023_02_01.models.KeyRotationPolicyActionType
     """
 
@@ -1271,7 +1274,8 @@ class ManagedHsmAction(_serialization.Model):
         self, *, type: Optional[Union[str, "_models.KeyRotationPolicyActionType"]] = None, **kwargs: Any
     ) -> None:
         """
-        :keyword type: The type of action. Known values are: "rotate" and "notify".
+        :keyword type: The type of the action. The value should be compared case-insensitively. Known
+         values are: "Rotate" and "Notify".
         :paramtype type: str or ~azure.mgmt.keyvault.v2023_02_01.models.KeyRotationPolicyActionType
         """
         super().__init__(**kwargs)
@@ -2019,7 +2023,7 @@ class ManagedHsmSku(_serialization.Model):
 
     All required parameters must be populated in order to send to Azure.
 
-    :ivar family: SKU Family of the managed HSM Pool. Required. "B"
+    :ivar family: SKU Family of the managed HSM Pool. "B"
     :vartype family: str or ~azure.mgmt.keyvault.v2023_02_01.models.ManagedHsmSkuFamily
     :ivar name: SKU of the managed HSM Pool. Required. Known values are: "Standard_B1",
      "Custom_B32", and "Custom_B6".
@@ -2039,12 +2043,12 @@ class ManagedHsmSku(_serialization.Model):
     def __init__(
         self,
         *,
-        family: Union[str, "_models.ManagedHsmSkuFamily"],
+        family: Union[str, "_models.ManagedHsmSkuFamily"] = "B",
         name: Union[str, "_models.ManagedHsmSkuName"],
         **kwargs: Any
     ) -> None:
         """
-        :keyword family: SKU Family of the managed HSM Pool. Required. "B"
+        :keyword family: SKU Family of the managed HSM Pool. "B"
         :paramtype family: str or ~azure.mgmt.keyvault.v2023_02_01.models.ManagedHsmSkuFamily
         :keyword name: SKU of the managed HSM Pool. Required. Known values are: "Standard_B1",
          "Custom_B32", and "Custom_B6".